# Conflicts:
#	src/main/kotlin/database/dao/AccountDAO.kt
#	src/main/kotlin/database/dao/DbTableMock.kt
#	src/main/kotlin/database/dao/UserInformationDAO.kt
#	src/main/kotlin/database/tables/UsersTable.kt
#	src/main/kotlin/domain/training/Training.kt
#	src/main/kotlin/domain/user/UserInfo.kt
#	src/test/kotlin/database/DbTableMockTest.kt

+++ src/main/kotlin/domain/training/Training.kt
@@ -47,7 +47,7 @@ sealed class Training(
     /**
      * Информация о беге
      */
-    class Jogging(date: LocalDate, duration: Duration, val distance: Double, id: Long? = null) : Training(
+    class Jogging(date: LocalDate, duration: Duration, val distance: Int, id: Long? = null) : Training(
         date = date,
         duration = duration,
         id = id
